splitting up train project to make things cleaner and documenting train function
