+++ projects/sandbox/pyproject.toml
@@ -80,7 +80,7 @@ force_generation = "${base.force_generation}"
 hanford_background = "${base.datadir}/H1_background.h5"
 livingston_background = "${base.datadir}/L1_background.h5"
 glitch_dataset = "${base.datadir}/glitches.h5"
-signal_dataset = "${base.datadir}/signals.h5"
+waveform_dataset = "${base.datadir}/signals.h5"
 logdir = "${base.logdir}"
 outdir = "${base.basedir}/training"
 
@@ -90,8 +90,8 @@ kernel_length = "${base.kernel_length}"
 fduration = "${base.fduration}"
 highpass = "${base.fmin}"
 trigger_distance = -0.5
-waveform_frac = 0.5
-glitch_frac = 0.25
+waveform_prob = 0.5
+glitch_prob = 0.25
 mean_snr = 10
 std_snr = 4
 min_snr = 2
